refactor: move SearchGroup into core
