+++ features/search/src/main/java/com/chesire/nekome/app/search/host/core/RememberSearchGroupUseCase.kt
@@ -1,7 +1,7 @@
 package com.chesire.nekome.app.search.host.core
 
+import com.chesire.nekome.app.search.host.core.model.SearchGroup
 import com.chesire.nekome.app.search.host.data.HostPreferences
-import com.chesire.nekome.app.search.host.domain.SearchGroup
 import javax.inject.Inject
 
 class RememberSearchGroupUseCase @Inject constructor(
