+++ features/search/src/main/java/com/chesire/nekome/app/search/host/core/model/SearchGroup.kt
@@ -1,0 +1,6 @@
+package com.chesire.nekome.app.search.host.core.model
+
+enum class SearchGroup {
+    Anime,
+    Manga
+}
