--- features/search/src/main/java/com/chesire/nekome/app/search/host/domain/SearchGroup.kt
@@ -1,6 +1,0 @@
-package com.chesire.nekome.app.search.host.domain
-
-enum class SearchGroup {
-    Anime,
-    Manga
-}
