+++ features/search/src/main/java/com/chesire/nekome/app/search/host/ui/UIState.kt
@@ -1,6 +1,6 @@
 package com.chesire.nekome.app.search.host.ui
 
-import com.chesire.nekome.app.search.host.domain.SearchGroup
+import com.chesire.nekome.app.search.host.core.model.SearchGroup
 
 data class UIState(
     val searchText: String,
